Main chunk of work for my dissertation project.
Created using Pycharm, this project includes a piece of software that allows users to attempt to scrape Twitter using snscrape. Due to the recent changes to the Twitter website, snscrape seemed to be having issues. Therefore, a static dataset was downloaded to showcase the capabilities of the software.

The software creates either 1 or multiple wordclouds (based on user settings). These wordclouds represent the highest-trending words associated with a hashtag inputted by the user. If the software creates multiple wordclouds, they are combined and displayed in an mp4 video, to show the change in trends over time.